requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
tensorflow==2.21.0
selenium==4.36.0
pillow==11.3.0
opencv-python==4.13.0.92
scikit-image==0.26.0
requests==2.32.5
numpy==2.3.3
pandas==2.3.2
scikit-learn==1.7.2
openpyxl==3.1.5
schedule==1.2.2
rapidfuzz==3.14.1
tldextract==5.2.0
beautifulsoup4==4.14.2
python-dateutil==2.9.0.post0
imagehash==4.3.2
pydantic==2.12.5
idna==3.11
